requirements.txt
python-dotenv
requests
reversinglabs-sdk-py3